+++ src/Netscoot.Core/Public/Sync-Solution.ps1
@@ -30,8 +30,8 @@ function Sync-Solution {
     [CmdletBinding(SupportsShouldProcess)]
     [OutputType('Netscoot.SyncResult')]
     param(
-        [Parameter(Position = 0, ValueFromPipeline, ValueFromPipelineByPropertyName)]
-        [Alias('FullName', 'Path', 'PSPath')]
+        [Parameter(Position = 0, ValueFromPipeline)]
+        [Netscoot.PathInputTransform()]
         [string]$RepositoryRoot
     )